server/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication & Security
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
django-oauth-toolkit==1.7.1

# Environment Configuration
python-decouple==3.8
django-environ==0.11.2

# Database
psycopg2-binary==2.9.7

# API Documentation
drf-spectacular==0.26.5
django-filter==23.3

# File Handling & Media
Pillow==10.0.1
django-storages==1.14.2
boto3==1.29.7

# Validation & Utilities
django-phonenumber-field==7.2.0
phonenumbers==8.13.24
django-extensions==3.2.3

# Caching & Performance
redis==4.5.4
django-redis==5.4.0
celery[redis]==5.3.4

# Testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0
coverage==7.3.2

# Development Tools
django-debug-toolbar==4.2.0
ipython==8.17.2
black==23.10.1
flake8==6.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Email
django-ses==3.5.0

# Task Queue
flower==2.0.1

# Social Media APIs
tweepy==4.14.0
facebook-sdk==3.1.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
instaloader==4.10.3

# Data Processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
httpx==0.25.2

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
urllib3<2.1.0,>=1.25.4  # Compatible with botocore requirements
